moved application to base; working server that generates a jpeg on startup
